Replace IntegrationKeyHandlerSegment implementation
