+++ app/src/userpages/components/ProfilePage/IdentityHandler/identityHandler.pcss
@@ -2,15 +2,6 @@
   margin-bottom: 2.25rem;
 }
 
-.keyList {
-  margin-bottom: 2rem;
-}
-
-.singleKey {
-  margin-top: 3rem;
-  margin-bottom: 2rem;
-}
-
 .addKey {
   margin-bottom: 2rem;
 }
