+++ app/src/userpages/components/ProfilePage/IdentityHandler/index.jsx
@@ -1,7 +1,7 @@
 // @flow
 
 import React, { Component, Fragment } from 'react'
-import { Row, Alert } from 'reactstrap'
+import { Alert } from 'reactstrap'
 import { connect } from 'react-redux'
 import { Translate } from 'react-redux-i18n'
 
@@ -10,7 +10,6 @@ import type { IntegrationKey } from '../../../flowtype/integration-key-types'
 import type { IntegrationKeyState } from '../../../flowtype/states/integration-key-state'
 import { deleteIntegrationKey, getIntegrationKeysByService, createIdentity } from '../../../modules/integrationKey/actions'
 import getWeb3 from '../../../utils/web3Provider'
-import KeyField from '$userpages/components/KeyField'
 
 import styles from './identityHandler.pcss'
 
@@ -55,34 +54,13 @@ export class IdentityHandler extends Component<Props> {
                 <div className={styles.description}>
                     <Translate value="userpages.profilePage.ethereumAddress.description" />
                 </div>
-                <div className={styles.keyList}>
-                    {this.props.integrationKeys.map((key: IntegrationKey) => (
-                        <KeyField
-                            className={styles.singleKey}
-                            key={key.id}
-                            keyName={key.name}
-                            // $FlowFixMe
-                            value={(key.json || {}).address || ''}
-                            allowDelete
-                            onDelete={() => {}}
-                        />
-                    ))}
-                </div>
-                <Row>
-                    <IntegrationKeyHandlerSegment
-                        onNew={this.onNew}
-                        onDelete={this.onDelete}
-                        service={service}
-                        integrationKeys={this.props.integrationKeys}
-                        copy="address"
-                        showInput={hasWeb3}
-                        tableFields={[
-                            ['address', (add) => (
-                                (add && typeof add === 'string') ? `${add.substring(0, 15)}...` : add
-                            )],
-                        ]}
-                    />
-                </Row>
+                <IntegrationKeyHandlerSegment
+                    onNew={this.onNew}
+                    onDelete={this.onDelete}
+                    service={service}
+                    integrationKeys={this.props.integrationKeys}
+                    showInput={hasWeb3}
+                />
                 {!hasWeb3 && (
                     <Alert color="danger">
                         To bind Ethereum addresses to your Streamr account, you need an Ethereum-enabled browser.
