--- app/src/userpages/components/ProfilePage/IntegrationKeyHandler/IntegrationKeyHandlerSegment/IntegrationKeyHandlerTable/IntegrationKeyHandlerTableRow/integrationKeyHandlerTableRow.pcss
@@ -1,21 +1,0 @@
-.deleteButton {
-  margin-right: -8px;
-}
-
-span.publicKey {
-  word-wrap: break-word;
-  max-width: 100%;
-}
-
-.actionButtonContainer {
-  display: flex;
-  justify-content: flex-end;
-
-  > * {
-    margin-left: 5px;
-
-    &:last-child {
-      margin-right: -8px;
-    }
-  }
-}
