--- app/src/userpages/components/ProfilePage/IntegrationKeyHandler/IntegrationKeyHandlerSegment/IntegrationKeyHandlerTable/integrationKeyHandlerTable.pcss
@@ -1,11 +1,0 @@
-.integrationKeyTable {
-  table-layout: fixed;
-
-  th.nameHeader {
-    width: 25%;
-  }
-
-  th.actionHeader {
-    width: 20%;
-  }
-}
