+++ app/src/userpages/components/ProfilePage/IntegrationKeyHandler/IntegrationKeyHandlerSegment/IntegrationKeyList/integrationKeyList.pcss
@@ -1,0 +1,8 @@
+.keyList {
+  margin-bottom: 2rem;
+}
+
+.singleKey {
+  margin-top: 3rem;
+  margin-bottom: 2rem;
+}
