+++ app/src/userpages/components/ProfilePage/IntegrationKeyHandler/IntegrationKeyHandlerSegment/index.jsx
@@ -1,23 +1,24 @@
 // @flow
 
-import React, { Component } from 'react'
-import { Col, Label } from 'reactstrap'
+import React, { Fragment, Component } from 'react'
+import { I18n } from 'react-redux-i18n'
+
 import type { IntegrationKey } from '../../../../flowtype/integration-key-types'
-import IntegrationKeyHandlerInput from './IntegrationKeyHandlerInput'
-import IntegrationKeyHandlerTable from './IntegrationKeyHandlerTable'
+import IntegrationKeyList from './IntegrationKeyList'
+import AddKeyField from '$userpages/components/KeyField/AddKeyField'
 
 import styles from './integrationKeyHandlerSegment.pcss'
 
-import type { Props as TableProps } from './IntegrationKeyHandlerTable'
-import type { Props as InputProps } from './IntegrationKeyHandlerInput'
+import type { Props as ListProps } from './IntegrationKeyList'
 
 type GivenProps = {
     className?: string,
     name?: $ElementType<IntegrationKey, 'name'>,
-    showInput: boolean
+    showInput: boolean,
+    onNew: (integrationKey: string) => void,
 }
 
-type Props = InputProps & TableProps & GivenProps
+type Props = ListProps & GivenProps
 
 export default class IntegrationKeyHandlerSegment extends Component<Props> {
     static defaultProps = {
@@ -26,27 +27,20 @@ export default class IntegrationKeyHandlerSegment extends Component<Props> {
 
     render() {
         return (
-            <div className={this.props.className || ''}>
-                <Col xs={12}>
-                    {this.props.name && (
-                        <Label className={styles.label}>
-                            {this.props.name}
-                        </Label>
-                    )}
-                    <IntegrationKeyHandlerTable
-                        tableFields={this.props.tableFields}
-                        integrationKeys={this.props.integrationKeys}
-                        onDelete={this.props.onDelete}
-                        copy={this.props.copy}
-                    />
-                    {this.props.showInput && (
-                        <IntegrationKeyHandlerInput
-                            inputFields={this.props.inputFields}
-                            onNew={this.props.onNew}
+            <Fragment>
+                <IntegrationKeyList
+                    integrationKeys={this.props.integrationKeys}
+                    onDelete={this.props.onDelete}
+                />
+                {this.props.showInput && (
+                    <div className={styles.addKey}>
+                        <AddKeyField
+                            label={I18n.t('userpages.profilePage.ethereumAddress.addNewAddress')}
+                            onSave={this.props.onNew}
                         />
-                    )}
-                </Col>
-            </div>
+                    </div>
+                )}
+            </Fragment>
         )
     }
 }
