+++ app/src/userpages/components/ProfilePage/IntegrationKeyHandler/IntegrationKeyHandlerSegment/integrationKeyHandlerSegment.pcss
@@ -1,3 +1,3 @@
-.label {
-  margin-top: 8px;
+.addKey {
+  margin-bottom: 2rem;
 }
